feat: share language selector between header and search
