+++ apps/watch/src/components/Header/Header.spec.tsx
@@ -1,75 +1,58 @@
-import { MockedProvider } from '@apollo/client/testing'
-import useScrollTrigger from '@mui/material/useScrollTrigger'
 import { fireEvent, render, screen } from '@testing-library/react'
 
 import { Header } from './Header'
 
-jest.mock('@mui/material/useScrollTrigger', () => ({
-  __esModule: true,
-  default: jest.fn()
+jest.mock('../SearchComponent/LanguageSelector', () => ({
+  LanguageSelector: () => <div data-testid="LanguageSelector" />
 }))
 
-const useScrollTriggerMock = useScrollTrigger as jest.Mock
+jest.mock('react-instantsearch', () => {
+  const actual = jest.requireActual('react-instantsearch')
+  return {
+    ...actual,
+    useInstantSearchContext: jest.fn()
+  }
+})
+
+const useInstantSearchContextMock = jest.requireMock('react-instantsearch')
+  .useInstantSearchContext as jest.Mock
 
 describe('Header', () => {
   beforeEach(() => {
-    useScrollTriggerMock.mockReset()
+    useInstantSearchContextMock.mockReset()
   })
 
-  it('should open navigation panel on menu icon click', async () => {
-    render(
-      <MockedProvider>
-        <Header />
-      </MockedProvider>
-    )
-    fireEvent.click(screen.getByRole('button', { name: 'open header menu' }))
-    expect(screen.getByRole('button', { name: 'Give' })).toBeInTheDocument()
-  })
+  it('should open navigation panel on menu icon click', () => {
+    useInstantSearchContextMock.mockImplementation(() => ({}))
 
-  it('should set menuOpen prop on LocalAppBar when menu is clicked', () => {
-    render(
-      <MockedProvider>
-        <Header />
-      </MockedProvider>
-    )
+    render(<Header />)
 
-    expect(
-      screen.queryByRole('button', { name: 'Give' })
-    ).not.toBeInTheDocument()
     const menuButton = screen.getByRole('button', { name: 'open header menu' })
     expect(menuButton).not.toHaveClass('expanded')
     expect(menuButton).toHaveAttribute('aria-expanded', 'false')
 
     fireEvent.click(menuButton)
+
     expect(menuButton).toHaveClass('expanded')
     expect(menuButton).toHaveAttribute('aria-expanded', 'true')
-    expect(screen.getByRole('button', { name: 'Give' })).toBeInTheDocument()
+    expect(screen.getByTestId('HeaderMenuPanel')).toBeInTheDocument()
   })
 
-  it('should hide bottom app bar', () => {
-    render(
-      <MockedProvider>
-        <Header hideBottomAppBar />
-      </MockedProvider>
-    )
-    expect(screen.queryByTestId('BottomAppBar')).not.toBeInTheDocument()
+  it('should render language selector when enabled inside instant search', () => {
+    useInstantSearchContextMock.mockImplementation(() => ({}))
+
+    render(<Header showLanguageSwitcher />)
+
+    expect(screen.getByTestId('LanguageSelector')).toBeInTheDocument()
   })
 
-  it('should hide spacer', () => {
-    render(
-      <MockedProvider>
-        <Header hideSpacer />
-      </MockedProvider>
-    )
-    expect(screen.queryByTestId('HeaderSpacer')).not.toBeInTheDocument()
-  })
+  it('should not render language selector without instant search context', () => {
+    useInstantSearchContextMock.mockImplementation(() => {
+      throw new Error('no context')
+    })
 
-  it('should hide top app bar', () => {
-    render(
-      <MockedProvider>
-        <Header hideTopAppBar />
-      </MockedProvider>
-    )
-    expect(screen.queryByTestId('TopAppBar')).not.toBeInTheDocument()
+    render(<Header showLanguageSwitcher />)
+
+    expect(screen.queryByTestId('LanguageSelector')).not.toBeInTheDocument()
   })
 })
