+++ apps/watch/src/components/Header/Header.tsx
@@ -1,12 +1,16 @@
 import Box from '@mui/material/Box'
 import Container from '@mui/material/Container'
+import IconButton from '@mui/material/IconButton'
 import SwipeableDrawer from '@mui/material/SwipeableDrawer'
+import MenuIcon from '@mui/icons-material/Menu'
 import { ReactElement, useState } from 'react'
+import { useInstantSearchContext } from 'react-instantsearch'
 
 import { ThemeProvider } from '@core/shared/ui/ThemeProvider'
 import { ThemeMode, ThemeName } from '@core/shared/ui/themes'
 
 import { HeaderMenuPanel } from './HeaderMenuPanel'
+import { LanguageSelector } from '../SearchComponent/LanguageSelector'
 
 /**
  * Props for the Header component.
@@ -15,6 +19,8 @@ import { HeaderMenuPanel } from './HeaderMenuPanel'
 interface HeaderProps {
   /** Theme mode to apply to the header. */
   themeMode?: ThemeMode
+  /** Display Algolia language filter in the header. */
+  showLanguageSwitcher?: boolean
 }
 
 /**
@@ -26,7 +32,8 @@ interface HeaderProps {
  * @returns {ReactElement} Rendered Header component.
  */
 export function Header({
-  themeMode = ThemeMode.light
+  themeMode = ThemeMode.light,
+  showLanguageSwitcher = false
 }: HeaderProps): ReactElement {
   /** State to control the drawer open/closed state. */
   const [drawerOpen, setDrawerOpen] = useState(false)
@@ -34,6 +41,42 @@ export function Header({
   return (
     <>
       <ThemeProvider themeName={ThemeName.website} themeMode={themeMode} nested>
+        <Box
+          component="header"
+          sx={{
+            width: '100%',
+            px: { xs: 4, md: 6 },
+            py: { xs: 3, md: 4 },
+            display: 'flex',
+            alignItems: 'center',
+            gap: 3
+          }}
+        >
+          <IconButton
+            aria-label="open header menu"
+            aria-expanded={drawerOpen}
+            className={drawerOpen ? 'expanded' : undefined}
+            onClick={() => setDrawerOpen(true)}
+            data-testid="MenuIcon"
+            size="large"
+            sx={{
+              color: 'text.primary',
+              border: '1px solid',
+              borderColor: 'divider',
+              borderRadius: '12px',
+              p: 1.5
+            }}
+          >
+            <MenuIcon />
+          </IconButton>
+          <Box sx={{ flex: 1, display: 'flex', justifyContent: 'flex-end' }}>
+            {showLanguageSwitcher === true && (
+              <Box sx={{ width: { xs: '100%', sm: 300 }, maxWidth: 360 }}>
+                <HeaderLanguageSelector />
+              </Box>
+            )}
+          </Box>
+        </Box>
       </ThemeProvider>
       <ThemeProvider
         themeName={ThemeName.website}
@@ -83,4 +126,14 @@ export function Header({
       </ThemeProvider>
     </>
   )
+}
+
+function HeaderLanguageSelector(): ReactElement | null {
+  try {
+    useInstantSearchContext()
+  } catch {
+    return null
+  }
+
+  return <LanguageSelector />
 }
