+++ libs/journeys/ui/src/components/LanguageSelector/index.ts
@@ -1,0 +1,1 @@
+export { LanguageSelector } from './LanguageSelector'
